# docs/openapi.yaml
openapi: 3.1.0
info:
  title: WordForge API
  description: |-
    WordForge 后端 API — v1-stable 端点。

    路径基础路径 `/api`（见 servers[0].url）。
    本规格由 `src/openapi.rs` 自动生成，通过 `cargo test --test openapi_export` 导出，CI 以 `git diff --exit-code docs/openapi.yaml` 防止规格漂移（drift）。
  version: 1.3.0-beta.1
servers:
- url: /api
paths:
...
  /word-states/batch:
    post:
      tags:
      - word-states
      summary: 批量查询单词学习状态
      description: 读侧批量查询（区别于 /word-states/batch-update 写侧）。数量上限见 config.rs limits.max_batch_size，超限 400 BATCH_TOO_LARGE。
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchQueryRequest'
        required: true
      responses:
        '200':
          description: WordLearningState 数组（每项附 bookmarked）
          content:
            application/json:
              schema:
                allOf:
                - type: object
                  required:
                  - success
                  properties:
                    success:
                      type: boolean
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      type: object
        '400':
          description: 请求体校验失败 — VALIDATION_ERROR / INVALID_REQUEST_BODY
        '401':
          description: 令牌缺失或已过期 — AUTH_UNAUTHORIZED
      security:
      - bearerAuth: []
  /word-states/batch-update:
    post:
      tags:
      - word-states
      summary: 批量更新单词学习状态
      description: 每项 state（lowercase：new/learning/reviewing/mastered/forgotten）/masteryLevel 均可选；只传 state 未传 masteryLevel 时，mastered/new 会强制对齐 masteryLevel 为 1.0/0.0。数量上限见 config.rs limits.max_batch_size。
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchUpdateRequest'
        required: true
      responses:
        '200':
          description: '`{updated: number}`，实际更新的条数'
          content:
            application/json:
              schema:
                allOf:
                - type: object
                  required:
                  - success
                  properties:
                    success:
                      type: boolean
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      type: object
        '400':
          description: 请求体校验失败 — VALIDATION_ERROR / INVALID_REQUEST_BODY
        '401':
          description: 令牌缺失或已过期 — AUTH_UNAUTHORIZED
      security:
      - bearerAuth: []
  /word-states/due/list:
    get:
      tags:
      - word-states
      summary: 获取当前已到期的复习单词列表
      description: 「已到期」含历史逾期（next_review_date <= now，非仅今日窗口）。limit 缺省 50，钳到 1..=200；到期总数超过 limit 时仅返回一页，真实总数看 /word-states/stats/overview 的 dueCount。
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
      responses:
        '200':
          description: 到期 WordLearningState 数组（每项附 bookmarked）
          content:
            application/json:
              schema:
                allOf:
                - type: object
                  required:
                  - success
                  properties:
                    success:
                      type: boolean
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      type: object
        '401':
          description: 令牌缺失或已过期 — AUTH_UNAUTHORIZED
      security:
      - bearerAuth: []
  /word-states/stats/overview:
    get:
      tags:
      - word-states
      summary: 单词学习状态统计总览
      description: category 缺省 all（可选 learning/review，非法值 400 INVALID_CATEGORY）。返回各状态计数 + dueCount / dueReviewEstimatedMinutes。dueCount 可空（序列化时 None 直接省略字段）：当前已到期总数（含历史逾期，未分页真实总量）——注意它恒为全局口径，不随 category 过滤变化。
      parameters:
      - name: category
        in: query
        required: false
        schema:
          type: string
          enum:
          - all
          - learning
          - review
      responses:
        '200':
          description: 统计总览：`{newCount, learning, reviewing, mastered, forgotten, dueReviewEstimatedMinutes?, dueCount?}`（后两者可空省略）
          content:
            application/json:
              schema:
                allOf:
                - type: object
                  required:
                  - success
                  properties:
                    success:
                      type: boolean
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      type: object
        '400':
          description: 请求体校验失败 — VALIDATION_ERROR / INVALID_REQUEST_BODY
        '401':
          description: 令牌缺失或已过期 — AUTH_UNAUTHORIZED
      security:
      - bearerAuth: []
  /word-states/{wordId}:
    get:
      tags:
      - word-states
      summary: 获取单词学习状态
      description: |-
        `state` 枚举值为 lowercase：`new/learning/reviewing/mastered/forgotten`
        （v0.6.0-beta.4 P3#7 起；与 MasteryLevel SCREAMING_SNAKE_CASE 是不同枚举）
      parameters:
      - name: wordId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: 单词学习状态（WordLearningState）
          content:
            application/json:
              schema:
                allOf:
                - type: object
                  required:
                  - success
                  properties:
                    success:
                      type: boolean
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      type: object
        '401':
          description: 令牌缺失或已过期 — AUTH_UNAUTHORIZED
        '404':
          description: 资源不存在 — NOT_FOUND
      security:
      - bearerAuth: []
  /word-states/{wordId}/mark-mastered:
    post:
      tags:
      - word-states
      summary: 标记单词为已掌握
      description: mastery_level 置 1.0，state 置 mastered，清除复习排程。
      parameters:
      - name: wordId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: 更新后的单词学习状态
          content:
            application/json:
              schema:
                allOf:
                - type: object
                  required:
                  - success
                  properties:
                    success:
                      type: boolean
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      type: object
        '401':
          description: 令牌缺失或已过期 — AUTH_UNAUTHORIZED
        '404':
          description: 资源不存在 — NOT_FOUND
      security:
      - bearerAuth: []
  /word-states/{wordId}/reset:
    post:
      tags:
      - word-states
      summary: 重置单词学习状态
      description: state 置 new，mastery_level/streak/attempts 全部清零，视为从未学过。
      parameters:
      - name: wordId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: 重置后的单词学习状态
          content:
            application/json:
              schema:
                allOf:
                - type: object
                  required:
                  - success
                  properties:
                    success:
                      type: boolean
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      type: object
        '401':
          description: 令牌缺失或已过期 — AUTH_UNAUTHORIZED
        '404':
          description: 资源不存在 — NOT_FOUND
      security:
      - bearerAuth: []
...
  /words/{wordId}:
    get:
      tags:
      - words
      summary: 获取单个单词详情
      parameters:
      - name: wordId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: 单词详情
          content:
            application/json:
              schema:
                allOf:
                - type: object
                  required:
                  - success
                  properties:
                    success:
                      type: boolean
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      type: object
        '401':
          description: 令牌缺失或已过期 — AUTH_UNAUTHORIZED
        '404':
          description: 资源不存在 — NOT_FOUND
      security:
      - bearerAuth: []
components:
  schemas:
...
    BatchQueryRequest:
      type: object
      required:
      - wordIds
      properties:
        wordIds:
          type: array
          items:
            type: string
    BatchUpdateRequest:
      type: object
      required:
      - updates
      properties:
        updates:
          type: array
          items:
            type: object
            required:
            - wordId
            properties:
              masteryLevel:
                type: number
              state:
                type: string
              wordId:
                type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...